Migrate a significant amount of StepMania.cpp into a new Main Project structure
StepManiaMain.cpp now is in control of bootstrap, display lifecycle, shutdown, and general main thread duties.  This changes no code,  only decouples these methods from StepMania.cpp so that StepMania.cpp can be scoped to only containing game facing utilities and input handling duties.
