+++ src/StepMania.cpp
@@ -84,246 +84,8 @@
 #include "UnlockManager.h"
 #include "ver.h"
 
-void ShutdownGame();
 bool HandleGlobalInputs(const InputEventPlus& input);
 void HandleInputEvents(float fDeltaTime);
-
-static Preference<bool> g_bAllowMultipleInstances(
-    "AllowMultipleInstances", false);
-
-void StepMania::GetPreferredVideoModeParams(VideoModeParams& paramsOut) {
-  // resolution handling code that probably needs fixing
-  int iWidth = PREFSMAN->m_iDisplayWidth;
-  if (PREFSMAN->m_bWindowed) {
-    // float fRatio = PREFSMAN->m_iDisplayHeight;
-    // iWidth = PREFSMAN->m_iDisplayHeight * fRatio;
-    iWidth =
-        std::ceil(PREFSMAN->m_iDisplayHeight * PREFSMAN->m_fDisplayAspectRatio);
-    // ceil causes the width to come out odd when it shouldn't.
-    // 576 * 1.7778 = 1024.0128, which is rounded to 1025. -Kyz
-    iWidth -= iWidth % 2;
-  }
-
-  paramsOut = VideoModeParams(
-      PREFSMAN->m_bWindowed || PREFSMAN->m_bFullscreenIsBorderlessWindow,
-      PREFSMAN->m_sDisplayId, iWidth, PREFSMAN->m_iDisplayHeight,
-      PREFSMAN->m_iDisplayColorDepth, PREFSMAN->m_iRefreshRate,
-      PREFSMAN->m_bVsync, PREFSMAN->m_bInterlaced, PREFSMAN->m_bSmoothLines,
-      PREFSMAN->m_bTrilinearFiltering, PREFSMAN->m_bAnisotropicFiltering,
-      !PREFSMAN->m_bWindowed && PREFSMAN->m_bFullscreenIsBorderlessWindow,
-      CommonMetrics::WINDOW_TITLE, THEME->GetPathG("Common", "window icon"),
-      PREFSMAN->m_bPAL, PREFSMAN->m_fDisplayAspectRatio);
-}
-
-static LocalizedString COLOR("StepMania", "color");
-static LocalizedString TEXTURE("StepMania", "texture");
-static LocalizedString WINDOWED("StepMania", "Windowed");
-static LocalizedString FULLSCREEN("StepMania", "Fullscreen");
-static LocalizedString ANNOUNCER_("StepMania", "Announcer");
-static LocalizedString VSYNC("StepMania", "Vsync");
-static LocalizedString NO_VSYNC("StepMania", "NoVsync");
-static LocalizedString SMOOTH_LINES("StepMania", "SmoothLines");
-static LocalizedString NO_SMOOTH_LINES("StepMania", "NoSmoothLines");
-
-static std::string GetActualGraphicOptionsString() {
-  const VideoModeParams& params = DISPLAY->GetActualVideoModeParams();
-  std::string sFormat = "%s %s %dx%d %d " + COLOR.GetValue() + " %d " +
-                        TEXTURE.GetValue() + " %dHz %s %s";
-  std::string sLog = ssprintf(
-      sFormat.c_str(), DISPLAY->GetApiDescription().c_str(),
-      (params.windowed ? WINDOWED : FULLSCREEN).GetValue().c_str(),
-      (int)params.width, (int)params.height, (int)params.bpp,
-      (int)PREFSMAN->m_iTextureColorDepth, (int)params.rate,
-      (params.vsync ? VSYNC : NO_VSYNC).GetValue().c_str(),
-      (PREFSMAN->m_bSmoothLines ? SMOOTH_LINES : NO_SMOOTH_LINES)
-          .GetValue()
-          .c_str());
-  return sLog;
-}
-
-static void StoreActualGraphicOptions() {
-  /* Store the settings that RageDisplay was actually able to use so that
-   * we don't go through the process of auto-detecting a usable video mode
-   * every time. */
-  const VideoModeParams& params = DISPLAY->GetActualVideoModeParams();
-  PREFSMAN->m_bWindowed.Set(
-      params.windowed && !params.bWindowIsFullscreenBorderless);
-  if (!params.windowed && !params.bWindowIsFullscreenBorderless) {
-    // In all other cases, want to preserve the value of this preference,
-    // but if DISPLAY decides to go fullscreen exclusive, we'll persist that
-    // decision
-    PREFSMAN->m_bFullscreenIsBorderlessWindow.Set(false);
-  }
-
-  /* If we're windowed, we may have tweaked the width based on the aspect ratio.
-   * Don't save this new value over the preferred value. */
-  if (!PREFSMAN->m_bWindowed) {
-    PREFSMAN->m_iDisplayWidth.Set(params.width);
-    PREFSMAN->m_iDisplayHeight.Set(params.height);
-  }
-  PREFSMAN->m_iDisplayColorDepth.Set(params.bpp);
-  if (PREFSMAN->m_iRefreshRate != REFRESH_DEFAULT) {
-    PREFSMAN->m_iRefreshRate.Set(params.rate);
-  }
-  PREFSMAN->m_bVsync.Set(params.vsync);
-
-  Dialog::SetWindowed(params.windowed);
-}
-
-static RageDisplay* CreateDisplay();
-
-bool StepMania::GetHighResolutionTextures() {
-  switch (PREFSMAN->m_HighResolutionTextures) {
-    default:
-    case HighResolutionTextures_Auto: {
-      int height = PREFSMAN->m_iDisplayHeight;
-      return height > THEME->GetMetricI("Common", "ScreenHeight");
-    }
-    case HighResolutionTextures_ForceOn:
-      return true;
-    case HighResolutionTextures_ForceOff:
-      return false;
-  }
-}
-
-static void update_centering() {
-  DISPLAY->ChangeCentering(
-      PREFSMAN->m_iCenterImageTranslateX, PREFSMAN->m_iCenterImageTranslateY,
-      PREFSMAN->m_fCenterImageAddWidth, PREFSMAN->m_fCenterImageAddHeight);
-}
-
-static void StartDisplay() {
-  if (DISPLAY != nullptr) {
-    return;  // already started
-  }
-
-  DISPLAY = CreateDisplay();
-
-  update_centering();
-
-  TEXTUREMAN = new RageTextureManager;
-  TEXTUREMAN->SetPrefs(RageTextureManagerPrefs(
-      PREFSMAN->m_iTextureColorDepth, PREFSMAN->m_iMovieColorDepth,
-      PREFSMAN->m_bDelayedTextureDelete, PREFSMAN->m_iMaxTextureResolution,
-      StepMania::GetHighResolutionTextures(), PREFSMAN->m_bForceMipMaps));
-
-  MODELMAN = new ModelManager;
-  MODELMAN->SetPrefs(ModelManagerPrefs(PREFSMAN->m_bDelayedModelDelete));
-}
-
-void StepMania::ApplyGraphicOptions() {
-  bool bNeedReload = false;
-
-  VideoModeParams params;
-  GetPreferredVideoModeParams(params);
-  std::string sError = DISPLAY->SetVideoMode(params, bNeedReload);
-  if (sError != "") {
-    RageException::Throw("%s", sError.c_str());
-  }
-
-  update_centering();
-
-  bNeedReload |= TEXTUREMAN->SetPrefs(RageTextureManagerPrefs(
-      PREFSMAN->m_iTextureColorDepth, PREFSMAN->m_iMovieColorDepth,
-      PREFSMAN->m_bDelayedTextureDelete, PREFSMAN->m_iMaxTextureResolution,
-      StepMania::GetHighResolutionTextures(), PREFSMAN->m_bForceMipMaps));
-
-  bNeedReload |=
-      MODELMAN->SetPrefs(ModelManagerPrefs(PREFSMAN->m_bDelayedModelDelete));
-
-  if (bNeedReload) {
-    TEXTUREMAN->ReloadAll();
-  }
-
-  StoreActualGraphicOptions();
-  if (SCREENMAN) {
-    SCREENMAN->SystemMessage(GetActualGraphicOptionsString());
-  }
-
-  // Give the input handlers a chance to re-open devices as necessary.
-  INPUTMAN->WindowReset();
-}
-
-static bool CheckVideoDefaultSettings();
-
-void StepMania::ResetPreferences() {
-  PREFSMAN->ResetToFactoryDefaults();
-  SOUNDMAN->SetMixVolume();
-  CheckVideoDefaultSettings();
-  ApplyGraphicOptions();
-}
-
-/* Shutdown all global singletons. Note that this may be called partway through
- * initialization, due to an object failing to initialize, in which case some of
- * these may still be nullptr. */
-void ShutdownGame() {
-  /* First, tell SOUNDMAN that we're shutting down. This signals sound drivers
-   * to stop sounds, which we want to do before any threads that may have
-   * started sounds are closed; this prevents annoying DirectSound glitches and
-   * delays. */
-  if (SOUNDMAN) {
-    SOUNDMAN->Shutdown();
-  }
-
-  /* Reset all lights to off.
-   * This is done before ~LightsManager as some drivers use SCREENMAN
-   * and similar when setting lights. */
-  if (LIGHTSMAN) {
-    LIGHTSMAN->TurnOffAllLights();
-  }
-
-  RageUtil::SafeDelete(NETWORK);
-  RageUtil::SafeDelete(SCREENMAN);
-  RageUtil::SafeDelete(STATSMAN);
-  RageUtil::SafeDelete(MESSAGEMAN);
-  /* Delete INPUTMAN before the other INPUTFILTER handlers, or an input
-   * driver may try to send a message to INPUTFILTER after we delete it. */
-  RageUtil::SafeDelete(INPUTMAN);
-  RageUtil::SafeDelete(INPUTQUEUE);
-  RageUtil::SafeDelete(INPUTMAPPER);
-  RageUtil::SafeDelete(INPUTFILTER);
-  RageUtil::SafeDelete(MODELMAN);
-  RageUtil::SafeDelete(PROFILEMAN);  // PROFILEMAN needs the songs still loaded
-  RageUtil::SafeDelete(CHARMAN);
-  RageUtil::SafeDelete(UNLOCKMAN);
-  RageUtil::SafeDelete(CRYPTMAN);
-  RageUtil::SafeDelete(MEMCARDMAN);
-  RageUtil::SafeDelete(SONGMAN);
-  RageUtil::SafeDelete(IMAGECACHE);
-  RageUtil::SafeDelete(SONGINDEX);
-  RageUtil::SafeDelete(SOUND);  // uses GAMESTATE, PREFSMAN
-  RageUtil::SafeDelete(PREFSMAN);
-  RageUtil::SafeDelete(GAMESTATE);
-  RageUtil::SafeDelete(GAMEMAN);
-  RageUtil::SafeDelete(NOTESKIN);
-  RageUtil::SafeDelete(THEME);
-  RageUtil::SafeDelete(ANNOUNCER);
-  RageUtil::SafeDelete(BOOKKEEPER);
-  RageUtil::SafeDelete(LIGHTSMAN);
-  RageUtil::SafeDelete(SOUNDMAN);
-  RageUtil::SafeDelete(FONT);
-  RageUtil::SafeDelete(TEXTUREMAN);
-  RageUtil::SafeDelete(DISPLAY);
-  Dialog::Shutdown();
-  RageUtil::SafeDelete(LUADEBUG);
-  RageUtil::SafeDelete(LOG);
-  RageUtil::SafeDelete(FILEMAN);
-  RageUtil::SafeDelete(LUA);
-  RageUtil::SafeDelete(HOOKS);
-}
-
-static void HandleException(const std::string& sError) {
-  if (g_bAutoRestart) {
-    HOOKS->RestartProgram();
-  }
-
-  // Shut down first, so we exit graphics mode before trying to open a dialog.
-  ShutdownGame();
-
-  // Throw up a pretty error dialog.
-  Dialog::Error(sError);
-  Dialog::Shutdown();  // Shut it back down.
-}
 
 void StepMania::ResetGame() {
   GAMESTATE->Reset();
@@ -356,273 +118,6 @@ std::string StepMania::GetInitialScreen() {
 ThemeMetric<std::string> SELECT_MUSIC_SCREEN("Common", "SelectMusicScreen");
 std::string StepMania::GetSelectMusicScreen() {
   return SELECT_MUSIC_SCREEN.GetValue();
-}
-
-#if defined(_WIN32)
-#include "RageDisplay_D3D.h"
-#include "archutils/Win32/VideoDriverInfo.h"
-#endif
-
-#if defined(SUPPORT_OPENGL)
-#include "RageDisplay_OGL.h"
-#endif
-
-#if defined(SUPPORT_GLES2)
-#include "RageDisplay_GLES2.h"
-#endif
-
-#include "RageDisplay_Null.h"
-
-struct VideoCardDefaults {
-  std::string sDriverRegex;
-  std::string sVideoRenderers;
-  int iWidth;
-  int iHeight;
-  int iDisplayColor;
-  int iTextureColor;
-  int iMovieColor;
-  int iTextureSize;
-  bool bSmoothLines;
-
-  VideoCardDefaults() {}
-  VideoCardDefaults(
-      std::string sDriverRegex_, std::string sVideoRenderers_, int iWidth_,
-      int iHeight_, int iDisplayColor_, int iTextureColor_, int iMovieColor_,
-      int iTextureSize_, bool bSmoothLines_) {
-    sDriverRegex = sDriverRegex_;
-    sVideoRenderers = sVideoRenderers_;
-    iWidth = iWidth_;
-    iHeight = iHeight_;
-    iDisplayColor = iDisplayColor_;
-    iTextureColor = iTextureColor_;
-    iMovieColor = iMovieColor_;
-    iTextureSize = iTextureSize_;
-    bSmoothLines = bSmoothLines_;
-  }
-} const g_VideoCardDefaults[] = {
-    // These lines correspond to the struct defined above.
-    VideoCardDefaults(
-        "OpenGL",   // Video card name (generic Mac/Linux)
-        "opengl",   // Available renderers
-        1280, 720,  // Default resolution
-        32,         // Display color
-        32,         // Texture color
-        32,         // Movie color
-        2048,       // Texture size
-        true        // Smooth lines
-        ),
-    VideoCardDefaults(
-        "",            // Video card name (generic Windows)
-        "opengl,d3d",  // Available renderers
-        1280, 720,     // Default resolution
-        32,            // Display color
-        32,            // Texture color
-        32,            // Movie color
-        2048,          // Texture size
-        true           // Smooth lines
-        ),
-};
-
-static std::string GetVideoDriverName() {
-#if defined(_WIN32)
-  return GetPrimaryVideoDriverName();
-#else
-  return "OpenGL";
-#endif
-}
-
-bool CheckVideoDefaultSettings() {
-  // Video card changed since last run
-  std::string sVideoDriver = GetVideoDriverName();
-
-  LOG->Trace(
-      "Last seen video driver: %s",
-      PREFSMAN->m_sLastSeenVideoDriver.Get().c_str());
-
-  VideoCardDefaults defaults;
-
-  unsigned i;
-  for (i = 0; i < ARRAYLEN(g_VideoCardDefaults); i++) {
-    defaults = g_VideoCardDefaults[i];
-
-    std::string sDriverRegex = defaults.sDriverRegex;
-    Regex regex(sDriverRegex);
-    if (regex.Compare(sVideoDriver)) {
-      LOG->Trace(
-          "Card matches '%s'.",
-          sDriverRegex.size() ? sDriverRegex.c_str() : "(unknown card)");
-      break;
-    }
-  }
-  if (i >= ARRAYLEN(g_VideoCardDefaults)) {
-    FAIL_M("Failed to match video driver");
-  }
-
-  bool bSetDefaultVideoParams = false;
-  if (PREFSMAN->m_sVideoRenderers.Get() == "") {
-    bSetDefaultVideoParams = true;
-    LOG->Trace("Applying defaults for %s.", sVideoDriver.c_str());
-  } else if (PREFSMAN->m_sLastSeenVideoDriver.Get() != sVideoDriver) {
-    bSetDefaultVideoParams = true;
-    LOG->Trace(
-        "Video card has changed from %s to %s.  Applying new defaults.",
-        PREFSMAN->m_sLastSeenVideoDriver.Get().c_str(), sVideoDriver.c_str());
-  }
-
-  if (bSetDefaultVideoParams) {
-    PREFSMAN->m_sVideoRenderers.Set(defaults.sVideoRenderers);
-    PREFSMAN->m_iDisplayWidth.Set(defaults.iWidth);
-    PREFSMAN->m_iDisplayHeight.Set(defaults.iHeight);
-    PREFSMAN->m_iDisplayColorDepth.Set(defaults.iDisplayColor);
-    PREFSMAN->m_iTextureColorDepth.Set(defaults.iTextureColor);
-    PREFSMAN->m_iMovieColorDepth.Set(defaults.iMovieColor);
-    PREFSMAN->m_iMaxTextureResolution.Set(defaults.iTextureSize);
-    PREFSMAN->m_bSmoothLines.Set(defaults.bSmoothLines);
-    // this only worked when we started in fullscreen by default. -aj
-    // PREFSMAN->m_fDisplayAspectRatio.Set( HOOKS->GetDisplayAspectRatio() );
-    // now that we start in windowed mode, use the new default aspect ratio.
-    PREFSMAN->m_fDisplayAspectRatio.Set(PREFSMAN->m_fDisplayAspectRatio);
-
-    // Update last seen video card
-    PREFSMAN->m_sLastSeenVideoDriver.Set(GetVideoDriverName());
-  } else if (CompareNoCase(
-                 PREFSMAN->m_sVideoRenderers.Get(), defaults.sVideoRenderers)) {
-    LOG->Warn(
-        "Video renderer list has been changed from '%s' to '%s'",
-        defaults.sVideoRenderers.c_str(),
-        PREFSMAN->m_sVideoRenderers.Get().c_str());
-  }
-
-  LOG->Info("Video renderers: '%s'", PREFSMAN->m_sVideoRenderers.Get().c_str());
-  return bSetDefaultVideoParams;
-}
-
-static LocalizedString ERROR_INITIALIZING_CARD(
-    "StepMania", "There was an error while initializing your video card.");
-static LocalizedString ERROR_DONT_FILE_BUG(
-    "StepMania",
-    "Please do not file this error as a bug!  Use the web page below to "
-    "troubleshoot this problem.");
-static LocalizedString ERROR_VIDEO_DRIVER("StepMania", "Video Driver: %s");
-static LocalizedString ERROR_NO_VIDEO_RENDERERS(
-    "StepMania", "No video renderers attempted.");
-static LocalizedString ERROR_INITIALIZING("StepMania", "Initializing %s...");
-static LocalizedString ERROR_UNKNOWN_VIDEO_RENDERER(
-    "StepMania", "Unknown video renderer value: %s");
-
-RageDisplay* CreateDisplay() {
-  /* We never want to bother users with having to decide which API to use.
-   *
-   * Some cards simply are too troublesome with OpenGL to ever use it, eg.
-   * Voodoos. If D3D8 isn't installed on those, complain and refuse to run (by
-   * default). For others, always use OpenGL.  Allow forcing to D3D as an
-   * advanced option.
-   *
-   * If we're missing acceleration when we load D3D8 due to a card being in the
-   * D3D list, it means we need drivers and that they do exist.
-   *
-   * If we try to load OpenGL and we're missing acceleration, it may mean:
-   *  1. We're missing drivers, and they just need upgrading.
-   *  2. The card doesn't have drivers, and it should be using D3D8.
-   *     In other words, it needs an entry in this table.
-   *  3. The card doesn't have drivers for either.  (Sorry, no S3 868s.)
-   *     Can't play.
-   * In this case, fail to load; don't silently fall back on D3D.  We don't want
-   * people unknowingly using D3D8 with old drivers (and reporting obscure bugs
-   * due to driver problems).  We'll probably get bug reports for all three
-   * types. #2 is the only case that's actually a bug.
-   *
-   * Actually, right now we're falling back. I'm not sure which behavior is
-   * better.
-   */
-
-  // bool bAppliedDefaults = CheckVideoDefaultSettings();
-  CheckVideoDefaultSettings();
-
-  VideoModeParams params;
-  StepMania::GetPreferredVideoModeParams(params);
-
-  std::string error =
-      ERROR_INITIALIZING_CARD.GetValue() + "\n\n" +
-      ERROR_DONT_FILE_BUG.GetValue() + "\n\n" VIDEO_TROUBLESHOOTING_URL "\n\n" +
-      ssprintf(
-          ERROR_VIDEO_DRIVER.GetValue().c_str(), GetVideoDriverName().c_str()) +
-      "\n\n";
-
-  std::vector<std::string> asRenderers;
-  split(PREFSMAN->m_sVideoRenderers, ",", asRenderers, true);
-
-  if (asRenderers.empty()) {
-    RageException::Throw("%s", ERROR_NO_VIDEO_RENDERERS.GetValue().c_str());
-  }
-
-  RageDisplay* pRet = nullptr;
-  for (unsigned i = 0; i < asRenderers.size(); i++) {
-    std::string sRenderer = asRenderers[i];
-
-    if (CompareNoCase(sRenderer, "opengl") == 0) {
-#if defined(SUPPORT_OPENGL)
-      pRet = new RageDisplay_Legacy;
-#endif
-    } else if (CompareNoCase(sRenderer, "gles2") == 0) {
-#if defined(SUPPORT_GLES2)
-      pRet = new RageDisplay_GLES2;
-#endif
-    } else if (CompareNoCase(sRenderer, "d3d") == 0) {
-// TODO: ANGLE/RageDisplay_Modern
-#if defined(SUPPORT_D3D)
-      pRet = new RageDisplay_D3D;
-#endif
-    } else if (CompareNoCase(sRenderer, "null") == 0) {
-      return new RageDisplay_Null;
-    } else {
-      RageException::Throw(
-          ERROR_UNKNOWN_VIDEO_RENDERER.GetValue().c_str(), sRenderer.c_str());
-    }
-
-    if (pRet == nullptr) {
-      continue;
-    }
-
-    std::string sError =
-        pRet->Init(params, PREFSMAN->m_bAllowUnacceleratedRenderer);
-    if (!sError.empty()) {
-      error +=
-          ssprintf(ERROR_INITIALIZING.GetValue().c_str(), sRenderer.c_str()) +
-          "\n" + sError;
-      RageUtil::SafeDelete(pRet);
-      error += "\n\n\n";
-      continue;
-    }
-
-    break;  // the display is ready to go
-  }
-
-  if (pRet == nullptr) {
-    RageException::Throw("%s", error.c_str());
-  }
-
-  return pRet;
-}
-
-static void SwitchToLastPlayedGame() {
-  const Game* pGame = GAMEMAN->StringToGame(PREFSMAN->GetCurrentGame());
-
-  // If the active game type isn't actually available, revert to the default.
-  if (pGame == nullptr) {
-    pGame = GAMEMAN->GetDefaultGame();
-  }
-
-  if (!GAMEMAN->IsGameEnabled(pGame) && pGame != GAMEMAN->GetDefaultGame()) {
-    pGame = GAMEMAN->GetDefaultGame();
-    LOG->Warn(
-        "Default NoteSkin for \"%s\" missing, reverting to \"%s\"",
-        pGame->m_szName, GAMEMAN->GetDefaultGame()->m_szName);
-  }
-
-  ASSERT(GAMEMAN->IsGameEnabled(pGame));
-
-  StepMania::InitializeCurrentGame(pGame);
 }
 
 // This function is meant to only be called during start up.
@@ -683,320 +178,6 @@ void StepMania::InitializeCurrentGame(const Game* g) {
   }
 }
 
-static void MountTreeOfZips(const std::string& dir) {
-  std::vector<std::string> dirs;
-  dirs.push_back(dir);
-
-  while (dirs.size()) {
-    std::string path = dirs.back();
-    dirs.pop_back();
-
-    if (!IsADirectory(path)) {
-      continue;
-    }
-
-    std::vector<std::string> zips;
-    GetDirListing(path + "/*.zip", zips, false, true);
-    GetDirListing(path + "/*.smzip", zips, false, true);
-
-    for (const auto& zip : zips) {
-      if (!IsAFile(zip)) {
-        continue;
-      }
-
-      LOG->Trace("VFS: found %s", zip.c_str());
-      FILEMAN->Mount("zip", zip, "/");
-    }
-
-    GetDirListing(path + "/*", dirs, true, true);
-  }
-}
-
-static void MountFolders(
-    const std::string& type, const std::string& realPathList,
-    const std::string& mountPoint) {
-  if (!realPathList.empty()) {
-    std::vector<std::string> dirs;
-    split(realPathList, ",", dirs, true);
-    for (const auto& dir : dirs) {
-      FILEMAN->Mount(type, dir, mountPoint);
-    }
-  }
-}
-
-static void WriteLogHeader() {
-  LOG->Info("%s%s", PRODUCT_FAMILY, product_version);
-
-  LOG->Info(
-      "Compiled %s @ %s (build %s)", version_date, version_time,
-      ::sm_version_git_hash);
-
-  time_t cur_time;
-  time(&cur_time);
-  struct tm now;
-  localtime_r(&cur_time, &now);
-
-  LOG->Info(
-      "Log starting %.4d-%.2d-%.2d %.2d:%.2d:%.2d", 1900 + now.tm_year,
-      now.tm_mon + 1, now.tm_mday, now.tm_hour, now.tm_min, now.tm_sec);
-  LOG->Trace(" ");
-
-  if (g_argc > 1) {
-    std::string args;
-    for (int i = 1; i < g_argc; ++i) {
-      if (i > 1) {
-        args += " ";
-      }
-
-      // surround all params with some marker, as they might have whitespace.
-      // using [[ and ]], as they are not likely to be in the params.
-      args += ssprintf("[[%s]]", g_argv[i]);
-    }
-    LOG->Info("Command line args (count=%d): %s", (g_argc - 1), args.c_str());
-  }
-}
-
-static void ApplyLogPreferences() {
-  LOG->SetShowLogOutput(PREFSMAN->m_bShowLogOutput);
-  LOG->SetLogToDisk(PREFSMAN->m_bLogToDisk);
-  LOG->SetInfoToDisk(PREFSMAN->m_bLogToDisk);
-  LOG->SetUserLogToDisk(PREFSMAN->m_bLogToDisk);
-  LOG->SetFlushing(PREFSMAN->m_bForceLogFlush);
-  Checkpoints::LogCheckpoints(PREFSMAN->m_bLogCheckpoints);
-}
-
-static LocalizedString COULDNT_OPEN_LOADING_WINDOW(
-    "LoadingWindow", "Couldn't open any loading windows.");
-
-int sm_main(int argc, char* argv[]) {
-  RageThreadRegister thread("Main thread");
-  RageException::SetCleanupHandler(HandleException);
-
-  SetCommandlineArguments(argc, argv);
-
-  // Set up arch hooks first.  This may set up crash handling.
-  HOOKS = ArchHooks::Create();
-  HOOKS->Init();
-
-  LUA = new LuaManager;
-  HOOKS->RegisterWithLua();
-
-  // Initialize the file extension type lists so everything can ask ActorUtil
-  // what the type of a file is.
-  ActorUtil::InitFileTypeLists();
-
-  // Almost everything uses this to read and write files.  Load this early.
-  FILEMAN = new RageFileManager(argv[0]);
-  FILEMAN->ProtectPath(SpecialFiles::DEFAULTS_INI_PATH);
-  FILEMAN->ProtectPath(SpecialFiles::STATIC_INI_PATH);
-  FILEMAN->ProtectPath(SpecialFiles::PREFERENCES_INI_PATH);
-  FILEMAN->MountInitialFilesystems();
-
-  bool bPortable = DoesFileExist("/Portable.ini");
-  if (!bPortable) {
-    FILEMAN->MountUserFilesystems();
-  }
-
-  // Set this up next. Do this early, since it's needed for
-  // RageException::Throw.
-  LOG = new RageLog;
-
-  // Whew--we should be able to crash safely now!
-
-  // load preferences and mount any alternative trees.
-  PREFSMAN = new PrefsManager;
-
-  /* Allow HOOKS to check for multiple instances.  We need to do this after
-   * PREFS is initialized, so ArchHooks can use a preference to turn this off.
-   * We want to do this before ApplyLogPreferences, so if we exit because of
-   * another instance, we don't try to clobber its log.  We also want to do this
-   * before opening the loading window, so if we give focus away, we don't flash
-   * the window. */
-  if (!g_bAllowMultipleInstances.Get() &&
-      HOOKS->CheckForMultipleInstances(argc, argv)) {
-    ShutdownGame();
-    return 0;
-  }
-
-  ApplyLogPreferences();
-
-  WriteLogHeader();
-
-  // Set up alternative filesystem trees.
-  MountFolders("dirro", PREFSMAN->m_sAdditionalFoldersReadOnly.Get(), "/");
-  MountFolders("dir", PREFSMAN->m_sAdditionalFoldersWritable.Get(), "/");
-  MountFolders(
-      "dirro", PREFSMAN->m_sAdditionalSongFoldersReadOnly.Get(), "/Songs");
-  MountFolders(
-      "dir", PREFSMAN->m_sAdditionalSongFoldersWritable.Get(), "/Songs");
-  MountFolders(
-      "dirro", PREFSMAN->m_sAdditionalCourseFoldersReadOnly.Get(), "/Courses");
-  MountFolders(
-      "dir", PREFSMAN->m_sAdditionalCourseFoldersWritable.Get(), "/Courses");
-
-  MountTreeOfZips(SpecialFiles::PACKAGES_DIR);
-
-  /* One of the above filesystems might contain files that affect preferences
-   * (e.g. Data/Static.ini). Re-read preferences. */
-  PREFSMAN->ReadPrefsFromDisk();
-  ApplyLogPreferences();
-
-  // This needs PREFSMAN.
-  Dialog::Init();
-
-  // Set up the messaging system early to have well defined code.
-  MESSAGEMAN = new MessageManager;
-
-  // Create game objects
-  GAMESTATE = new GameState;
-
-  // This requires PREFSMAN, for PREFSMAN->m_bShowLoadingWindow.
-  LoadingWindow* pLoadingWindow = LoadingWindow::Create();
-  if (pLoadingWindow == nullptr) {
-    RageException::Throw("%s", COULDNT_OPEN_LOADING_WINDOW.GetValue().c_str());
-  }
-
-  srand(time(nullptr));  // seed number generator
-
-  /* Do this early, so we have debugging output if anything else fails. LOG and
-   * Dialog must be set up first. It shouldn't take long, but it might take a
-   * little time; do this after the LoadingWindow is shown, since we don't want
-   * that to appear delayed. */
-  HOOKS->DumpDebugInfo();
-
-#if defined(HAVE_TLS)
-  LOG->Info("TLS is %savailable", RageThread::GetSupportsTLS() ? "" : "not ");
-#endif
-
-  std::string luaDebugServerAddress;
-  if (GetCommandlineArgument("lua-debugger", &luaDebugServerAddress)) {
-    LUADEBUG = new LuaDebugManager();
-    LUADEBUG->Start(
-        luaDebugServerAddress, GetCommandlineArgument("lua-debugger-paused"));
-  }
-
-  GAMEMAN = new GameManager;
-  THEME = new ThemeManager;
-  ANNOUNCER = new AnnouncerManager;
-  NOTESKIN = new NoteSkinManager;
-
-  // Switch to the last used game type, and set up the theme and announcer.
-  SwitchToLastPlayedGame();
-
-  CommandLineActions::Handle(pLoadingWindow);
-
-  if (GetCommandlineArgument("dopefish")) {
-    GAMESTATE->m_bDopefish = true;
-  }
-
-  {
-    /* Now that THEME is loaded, load the icon and splash for the current
-     * theme into the loading window. */
-    std::string sError;
-    RageSurface* pSurface = RageSurfaceUtils::LoadFile(
-        THEME->GetPathG("Common", "window icon"), sError);
-    if (pSurface != nullptr) {
-      pLoadingWindow->SetIcon(pSurface);
-    }
-    delete pSurface;
-    pSurface =
-        RageSurfaceUtils::LoadFile(THEME->GetPathG("Common", "splash"), sError);
-    if (pSurface != nullptr) {
-      pLoadingWindow->SetSplash(pSurface);
-    }
-    delete pSurface;
-  }
-
-  if (PREFSMAN->m_iSoundWriteAhead) {
-    LOG->Info(
-        "Sound writeahead has been overridden to %i",
-        PREFSMAN->m_iSoundWriteAhead.Get());
-  }
-
-  SOUNDMAN = new RageSoundManager;
-  SOUNDMAN->Init();
-  SOUNDMAN->SetMixVolume();
-  SOUND = new GameSoundManager;
-  BOOKKEEPER = new Bookkeeper;
-  LIGHTSMAN = new LightsManager;
-  INPUTFILTER = new InputFilter;
-  INPUTMAPPER = new InputMapper;
-
-  StepMania::InitializeCurrentGame(GAMESTATE->GetCurrentGame());
-
-  INPUTQUEUE = new InputQueue;
-  SONGINDEX = new SongCacheIndex;
-  IMAGECACHE = new ImageCache;
-
-  // depends on SONGINDEX:
-  SONGMAN = new SongManager;
-  SONGMAN->InitAll(
-      pLoadingWindow, /*onlyAdditions=*/false);  // this takes a long time
-  CRYPTMAN = new CryptManager;  // need to do this before ProfileMan
-  if (PREFSMAN->m_bSignProfileData) {
-    CRYPTMAN->GenerateGlobalKeys();
-  }
-  MEMCARDMAN = new MemoryCardManager;
-  CHARMAN = new CharacterManager;
-  PROFILEMAN = new ProfileManager;
-  PROFILEMAN->Init();  // must load after SONGMAN
-  UNLOCKMAN = new UnlockManager;
-  SONGMAN->UpdatePopular();
-  SONGMAN->UpdatePreferredSort();
-  NETWORK = new NetworkManager;
-  STATSMAN = new StatsManager;
-
-  // Initialize which courses are ranking courses here.
-  SONGMAN->UpdateRankingCourses();
-
-  RageUtil::SafeDelete(pLoadingWindow);  // destroy this before init'ing Display
-
-  /* If the user has tried to quit during the loading, do it before creating
-   * the main window. This prevents going to full screen just to quit. */
-  if (ArchHooks::UserQuit()) {
-    ShutdownGame();
-    return 0;
-  }
-
-  StartDisplay();
-
-  StoreActualGraphicOptions();
-  LOG->Info("%s", GetActualGraphicOptionsString().c_str());
-
-  SONGMAN->PreloadSongImages();
-
-  /* Input handlers can have dependences on the video system so
-   * INPUTMAN must be initialized after DISPLAY. */
-  INPUTMAN = new RageInput;
-
-  // These things depend on the TextureManager, so do them after!
-  FONT = new FontManager;
-  SCREENMAN = new ScreenManager;
-
-  StepMania::ResetGame();
-
-  /* Now that GAMESTATE is reset, tell SCREENMAN to update the theme (load
-   * overlay screens and global sounds), and load the initial screen. */
-  SCREENMAN->ThemeChanged();
-  SCREENMAN->SetNewScreen(StepMania::GetInitialScreen());
-
-  // Do this after ThemeChanged so that we can show a system message
-  std::string sMessage;
-  if (INPUTMAPPER->CheckForChangedInputDevicesAndRemap(sMessage)) {
-    SCREENMAN->SystemMessage(sMessage);
-  }
-
-  CodeDetector::RefreshCacheItems();
-
-  // Run the main loop.
-  GameLoop::RunGameLoop();
-
-  PREFSMAN->SavePrefsToDisk();
-
-  ShutdownGame();
-
-  return 0;
-}
 
 std::string StepMania::SaveScreenshot(
     std::string Dir, bool SaveCompressed, bool MakeSignature,
@@ -1430,12 +611,6 @@ void LuaFunc_Register_SaveScreenshot(lua_State* L) {
   lua_register(L, "SaveScreenshot", LuaFunc_SaveScreenshot);
 }
 REGISTER_WITH_LUA_FUNCTION(LuaFunc_Register_SaveScreenshot);
-
-static int LuaFunc_update_centering(lua_State* L) {
-  update_centering();
-  return 0;
-}
-LUAFUNC_REGISTER_COMMON(update_centering);
 
 /*
  * (c) 2001-2004 Chris Danford, Glenn Maynard
